+++ MercadoPagoSDK/MercadoPagoSDK/Flows/OneTap/UI/PXOneTapSplitPaymentView.swift
@@ -12,6 +12,8 @@ class PXOneTapSplitPaymentView: PXComponentView {
     var splitConfiguration: PXSplitConfiguration?
     var splitPaymentSwitch: UISwitch?
     var splitMessageLabel: UILabel?
+    var separatorView: UIView?
+    private let switchReduceSize: CGFloat = 0.70
 
     init(splitConfiguration: PXSplitConfiguration?, callback : @escaping ((_ isOn: Bool, _ isUserSelection: Bool) -> Void)) {
         self.splitConfiguration = splitConfiguration
@@ -27,11 +29,13 @@ class PXOneTapSplitPaymentView: PXComponentView {
     func hide() {
         splitMessageLabel?.alpha = 0
         splitPaymentSwitch?.alpha = 0
+        separatorView?.isHidden = true
     }
 
     func show() {
         splitMessageLabel?.alpha = 1
         splitPaymentSwitch?.alpha = 1
+        separatorView?.isHidden = false
     }
 
     func update(splitConfiguration: PXSplitConfiguration?) {
@@ -63,18 +67,21 @@ class PXOneTapSplitPaymentView: PXComponentView {
 
         splitSwitch.setOn(splitConfiguration?.splitEnabled ?? false, animated: false)
         splitSwitch.translatesAutoresizingMaskIntoConstraints = false
+        splitSwitch.transform = CGAffineTransform(scaleX: switchReduceSize, y: switchReduceSize)
         self.addSubview(splitSwitch)
-        PXLayout.pinRight(view: splitSwitch, withMargin: PXLayout.L_MARGIN).isActive = true
+        PXLayout.pinRight(view: splitSwitch, withMargin: PXLayout.M_MARGIN).isActive = true
         PXLayout.centerVertically(view: splitSwitch).isActive = true
         splitSwitch.onTintColor = ThemeManager.shared.getAccentColor()
 
         let label = UILabel()
+        label.lineBreakMode = .byTruncatingTail
         self.splitMessageLabel = label
         label.translatesAutoresizingMaskIntoConstraints = false
+        label.textAlignment = .left
         self.addSubview(label)
         PXLayout.centerVertically(view: label).isActive = true
         PXLayout.pinLeft(view: label, withMargin: PXLayout.L_MARGIN).isActive = true
-        PXLayout.put(view: label, leftOf: splitSwitch, withMargin: PXLayout.XXXS_MARGIN).isActive = true
+        PXLayout.pinRight(view: label, to: splitSwitch, withMargin: PXLayout.XXXL_MARGIN).isActive = true
         PXLayout.pinTop(view: label, withMargin: PXLayout.S_MARGIN).isActive = true
         PXLayout.pinBottom(view: label, withMargin: PXLayout.S_MARGIN).isActive = true
 
@@ -86,6 +93,7 @@ class PXOneTapSplitPaymentView: PXComponentView {
 
         let separatorView = UIView()
         separatorView.translatesAutoresizingMaskIntoConstraints = false
+        self.separatorView = separatorView
         separatorView.backgroundColor = UIColor.black.withAlphaComponent(0.1)
         self.addSubview(separatorView)
         PXLayout.setHeight(owner: separatorView, height: 1).isActive = true
@@ -109,9 +117,13 @@ class PXOneTapSplitPaymentView: PXComponentView {
 
         let attributes: [NSAttributedString.Key: AnyObject] = [NSAttributedString.Key.font: Utils.getSemiBoldFont(size: PXLayout.XS_FONT), NSAttributedString.Key.foregroundColor: ThemeManager.shared.boldLabelTintColor()]
 
+        let messageAttributes: [NSAttributedString.Key: AnyObject] = [NSAttributedString.Key.font: Utils.getFont(size: PXLayout.XS_FONT), NSAttributedString.Key.foregroundColor: ThemeManager.shared.greyColor()]
+
+        let messageAttributed = NSAttributedString(string: splitConfiguration.secondaryPaymentMethod?.message ?? "", attributes: messageAttributes)
+
         let amountAttributed = NSMutableAttributedString(string: amount, attributes: attributes)
         amountAttributed.append(" ".toAttributedString())
-        amountAttributed.append(splitConfiguration.secondaryPaymentMethod?.message?.toAttributedString() ?? "".toAttributedString())
+        amountAttributed.append(messageAttributed)
         return amountAttributed
     }